Linked Reviews and Cart to USERS in backend Made it so frontend now triggers these endpoints and populates DB Some UI changes as well

+++ bookstore-backend/src/main/java/com/bookstore/backend/Book/BookController.java
@@ -34,7 +34,7 @@ public class BookController {
 
     // Get book by ID
     @GetMapping("/{id}")
-    public ResponseEntity<Book> getBookById(@PathVariable Long id) {
+    public ResponseEntity<Book> getBookById(@PathVariable int id) {
         return bookRepository.findById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
